DRT-Explorer(server)_24_07_16/requirements.txt
python-socketio==4.6.0
python-engineio==3.13.2
sanic==20.6.3
matplotlib==3.4.1